Client: Moved FriendShareList into own friend-system feature
