+++ client/src/features/friendship-system/index.js
@@ -1,0 +1,5 @@
+export { FriendShareList } from "./components/FriendShareList";
+
+export { acceptFriendRequest } from "./services/friendRequestFetches";
+export { rejectFriendRequest } from "./services/friendRequestFetches";
+export { sendFriendRequest } from "./services/friendRequestFetches";
