+++ client/src/features/message-center/index.js
@@ -1,5 +1,1 @@
-export { MessageCenter } from "./pages/MessageCenter";
-
-export { acceptFriendRequest } from "./services/friendRequestFetches.js";
-export { rejectFriendRequest } from "./services/friendRequestFetches.js";
-export { sendFriendRequest } from "./services/friendRequestFetches.js";
+export { MessageCenter } from "../../pages/MessageCenter";
